REF: Move energy test functions to new .drivers module
