+++ .github/workflows/ci.yaml
@@ -76,7 +76,7 @@ jobs:
       continue-on-error: true
       shell: bash -l {0}
       run: |
-        mypy --show-error-codes --warn-unused-configs --warn-unused-ignored --namespace-packages -p "openff.system"
+        mypy --show-error-codes --warn-unused-configs --warn-unused-ignores --namespace-packages -p "openff.system"
 
     - name: Run unit tests
       if: always()
