+++ openff/system/tests/integration_tests/test_interoperability_pathways.py
@@ -8,8 +8,8 @@ from openff.units import unit
 from openff.utilities import temporary_cd
 from simtk import unit as omm_unit
 
+from openff.system.drivers.gromacs import _get_mdp_file, _run_gmx_energy
 from openff.system.stubs import ForceField
-from openff.system.tests.energy_tests.gromacs import _get_mdp_file, _run_gmx_energy
 from openff.system.tests.energy_tests.test_energies import needs_gmx
 from openff.system.types import ArrayQuantity
